Moved the code into AlbumPage component to get the code ready for being able to toggle different code with a button. Not sure if I need to do this but all the examples use classes and the uploads is a const in index.js.
