+++ src/assignments/project-1/index.js
@@ -1,10 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import Uploader from '../../ui/components/uploader'
-import SectionHeader from './sectionheader'
-import Updater from './updater'
-import Album from './album'
-import AlbumDescription from './albumdescription'
+import AlbumPage from './albumpage'
 
 import './app.scss'
 
@@ -15,21 +11,8 @@ const Uploads = ({uploads, actions}) => {
 	return <div>
 		{/* Link in Desired Font Families */}
 		<link rel="stylesheet" type="text/css" href="//fonts.googleapis.com/css?family=Roboto"/>
-
-		<SectionHeader tag="h1" sectiontext="My Project 1 Album"/>
-		<AlbumDescription paragraphs={[FIRSTPARAGRAPH, SECONDPARAGRAPH]} />
-		<SectionHeader tag="h1" sectiontext="Upload Images"/>
-		{/* do not delete this uploader component */}
-		<Uploader className="file-form" upload={actions.upload} />
-		{/* do not delete this uploader component */}
-
-		{/* Display PendingFiles Header and List */}
-		<SectionHeader tag="h2" sectiontext="In Progress"/>
-		<Album isPendingFiles={true} photos={pendingFiles} updateFile={actions.files.updateFile}/>
-
-		{/* Display CompletedFiles Header and List */}
-		<SectionHeader tag="h2" sectiontext="Completed"/>
-		<Album isPendingFiles={false} photos={completedFiles} updateFile={actions.files.updateFile}/>
+		<AlbumPage pendingFiles={pendingFiles} completedFiles={completedFiles} actions={actions}/>
+		
 	</div>
 }
 
@@ -53,21 +36,6 @@ Uploads.propTypes = {
 		share: statusPropType.isRequired,
 	}).isRequired,
 	actions: PropTypes.object.isRequired,
-}
-
-
-const FIRSTPARAGRAPH = {
-	header: "This is my content header...",
-	header_class: "content_header",
-	text: "This album is my first attempt at using Reactjs. As you will probably see, I still need lots of help with React. I will attempt to pass this paragraph data from many items into the uploader index file. I will also pass in some simple class information into the react component.",
-	text_class: "paragraph_text",
-}
-
-const SECONDPARAGRAPH = {
-	header: "This is the second header content...",
-	header_class: "content_header",
-	text: "This is the second paragraph that I am attempting to write. The images that I am using for this shown below are randomly found on my local computer so they might not make any sense.",
-	text_class: "paragraph_text",
 }
 
 export default Uploads
